fix: Click on several pairs
Pairs can now be clicked, it was limited to one.
Now you can almost finish the game.
I still have to disable the double click on the same image.
I have to remove the umatched class after a few seconds of display
I also have to add some endGame rules because the game did not end
when I clicked on every pairs. It even allowed me to go up to -5.